Made the login click a listener and removed it from the public API.
Now none of Foamicator is exposed to other addons. This helps avoid
namespace collisions and also provides better security because another
addon can't just access the variables.

Also moved the encryption key to memory instead of the password manager.
The master password is verified using a hash that is stored in the
database.